refactor: extract startup tls material boundary
